Query games more robustly
There are some quirks around how Rust and Redis handle the distinction
between GET and MGET. There should always be at least one element in the
requested key list and the redis-rs implementation also only uses MGET
if more than one key is requested. This is irrespective of the fact if a
Vec was used as a parameter.